update link script

+++ black_box/link_dotfiles/link.sh
@@ -3,12 +3,12 @@
 DOTFILES_DIR=../../dotfiles/
 CONFIG_DIR=~/.config/
 
-if [ ! -d $DOTFILES_DIR ]; then
+if [ ! -d "$DOTFILES_DIR" ]; then
     echo "Error: $DOTFILES_DIR does not exist."
     exit 1
 fi
 
-if [ ! -d $CONFIG_DIR ]; then
+if [ ! -d "$CONFIG_DIR" ]; then
     echo "Error: $CONFIG_DIR does not exist."
     exit 1
 fi
@@ -22,10 +22,17 @@ for item in "$DOTFILES_DIR"/*; do
     if [[ -e "$target" ]]; then
         echo "Skipping $name: Target already exists."
     else
-        echo "Creating symlink for $name..."
-        ln -s "$item" "$target"
+        if [[ -d "$item" ]]; then
+            echo "Creating recursive symlink for directory $name..."
+            ln -s -r "$item" "$target"
+        elif [[ -f "$item" ]]; then
+            echo "Creating symlink for file $name..."
+            ln -s "$item" "$target"
+        else
+            echo "Skipping $name: Not a regular file or directory."
+            continue
+        fi
     fi
-
 done
 
 echo "All files and directories symlinked successfully!"
